notebooks/requirements.txt
ipykernel==6.18.1
ipython==8.7.0
Pillow==9.3.0
python-dotenv==0.21.0
requests==2.28.1
plantcv==3.14.2
matplotlib==3.5.3